Web-site - added checking image for null before deleting in service page

+++ my-house24-admin/src/main/java/com/example/myhouse24admin/serviceImpl/ServicesPageServiceImpl.java
@@ -109,9 +109,14 @@ public class ServicesPageServiceImpl implements ServicesPageService {
         if(image.isEmpty()){
             return servicePageBlockInDb.getImage();
         } else {
-            uploadFileUtil.deleteFile(servicePageBlockInDb.getImage());
+            if(imageNotNull(servicePageBlockInDb)) {
+                uploadFileUtil.deleteFile(servicePageBlockInDb.getImage());
+            }
             return uploadFileUtil.saveMultipartFile(image);
         }
+    }
+    private boolean imageNotNull(ServicePageBlock servicePageBlockInDb){
+        return servicePageBlockInDb.getImage() != null;
     }
 
 }
